xinput-calibrator: Change wrapper script to allow it to be run by a normal user
This is a backport from Daisy (minus moving the recipe from meta-oe) of:
commit 4ecafd89dbf41cbd53e6db32678fe625c06caaab

    xinput-calibrator: move it from meta-oe to oe-core

    People using xserver-xorg that need to calibrate their touchscreen
    devices would also need meta-oe. Bringing the recipes to oe-core will
    make it easier for them.

    Aditionaly:
     * drop xterm RDEPENDS. Terminal is not needed to run the menu item;
     * change xinput_calibrator_pointercal.sh so that it can be run as
       normal user: pointercal.xinput is saved to ~/.pointercal/ and it will
       be used when the system boots;
     * have the calibration run through an Xsession file;
     * remove the systemd service since calibration is run by Xsession;
     * do not install pointercal.xinput if it's the default one;

    [YOCTO #4416]


Acked-by: Ken Sharp <ken.sharp@ni.com>
---
Natinst-ReviewBoard-ID: 80151

+++ meta-oe/recipes-graphics/xinput-calibrator/pointercal-xinput_0.0.bb
@@ -9,7 +9,8 @@ S = "${WORKDIR}"
 
 do_install() {
     # Only install file if it has a contents
-    if [ -s ${S}/pointercal.xinput ]; then
+    if [ -s ${S}/pointercal.xinput ] &&\
+       [ ! -n "$(head -n1 ${S}/pointercal.xinput|grep "replace.*pointercal\.xinput")" ]; then
         install -d ${D}${sysconfdir}/
         install -m 0644 ${S}/pointercal.xinput ${D}${sysconfdir}/
     fi
